+++ internal/generator/generator_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"gitlab.wildberries.ru/travel/travel-backend/adv/core/hz-openapi-gen/internal/diag"
-	"gitlab.wildberries.ru/travel/travel-backend/adv/core/hz-openapi-gen/internal/openapi"
+	"github.com/victorzhuk/hz-openapi-gen/internal/diag"
+	"github.com/victorzhuk/hz-openapi-gen/internal/openapi"
 )
 
 func generateMinimal(t *testing.T) []GeneratedFile {