+++ internal/writer/writer_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"gitlab.wildberries.ru/travel/travel-backend/adv/core/hz-openapi-gen/internal/generator"
+	"github.com/victorzhuk/hz-openapi-gen/internal/generator"
 )
 
 func marked(path, body string) generator.GeneratedFile {