+++ testdata/golden/generate.go
@@ -2,6 +2,6 @@
 
 package main
 
-//go:generate go run gitlab.wildberries.ru/travel/travel-backend/adv/core/hz-openapi-gen/cmd/hz-openapi-gen -spec=testdata/minimal.yaml -out=. -module=example.com/service
+//go:generate go run github.com/victorzhuk/hz-openapi-gen/cmd/hz-openapi-gen -spec=testdata/minimal.yaml -out=. -module=example.com/service
 
 func main() {}